AbstractingClass commit

+++ src/main/java/com/core/timmy/data/repository/IBudgetRepository.java
@@ -11,6 +11,6 @@ import com.core.timmy.data.model.Status;
 //para que identifique esto como un repository le hacemos la anotación así sabe que es una entity o entidad
 
 @Repository
-public interface IBudgetRepository extends JpaRepository<Budget, Integer> {
+public interface IBudgetRepository extends JpaRepository<Budget, Long> {
 
 }
